move parts of the text editor changes to the file service
