+++ src/components/fileService.ts
@@ -1,10 +1,13 @@
-import { FilePermission, TextDocument, Uri } from "vscode";
+import { FilePermission, Position, Range, Selection, TextDocument, TextEditor, TextEditorRevealType, Uri, window, workspace } from "vscode";
+import { Configuration } from "../configuration";
+import { Container } from "../container";
 import { destructIntuitaFileSystemUri } from "../destructIntuitaFileSystemUri";
 import { JobManager } from "./jobManager";
 import { Message, MessageBus, MessageKind } from "./messageBus";
 
 export class FileService {
     public constructor(
+        protected readonly _configurationContainer: Container<Configuration>,
         protected readonly _jobManager: JobManager,
         protected readonly _messageBus: MessageBus,
         protected readonly _openTextDocument: (uri: Uri) => Promise<TextDocument> 
@@ -12,6 +15,10 @@ export class FileService {
         this._messageBus.subscribe(async (message) => {
             if (message.kind === MessageKind.readingFileFailed) {
                 setImmediate(() => this._onReadingFileFailed(message));
+            }
+
+            if (message.kind === MessageKind.updateExternalFile) {
+                setImmediate(() => this._onUpdateExternalFile(message));
             }
         });
     }
@@ -57,5 +64,71 @@ export class FileService {
         const textDocument = await this._openTextDocument(uri);
 
         return textDocument.getText();        
+    }
+
+    protected async _onUpdateExternalFile(
+        message: Message & { kind: MessageKind.updateExternalFile },
+    ) {
+        const stringUri = message.uri.toString();
+
+        const textEditors = window.visibleTextEditors.filter(({ document }) => {
+			return document.uri.toString() === stringUri;
+		});
+
+		const textDocuments = workspace.textDocuments.filter((document) => {
+			return document.uri.toString() === stringUri;
+		});
+
+		const activeTextEditor = window.activeTextEditor ?? null;
+
+		const range = new Range(
+			new Position(message.jobOutput.range[0], message.jobOutput.range[1]),
+			new Position(message.jobOutput.range[2], message.jobOutput.range[3]),
+		);
+
+		const { saveDocumentOnJobAccept } = this._configurationContainer.get();
+
+		const changeTextEditor = async (textEditor: TextEditor) => {
+			await textEditor.edit((textEditorEdit) => {
+				textEditorEdit.replace(range, message.jobOutput.text);
+			});
+
+			if (!saveDocumentOnJobAccept) {
+				return;
+			}
+
+			return textEditor.document.save();
+		};
+
+		await Promise.all(textEditors.map(changeTextEditor));
+
+		if (textEditors.length === 0) {
+			textDocuments.forEach((textDocument) => {
+				window
+					// TODO we can add a range here
+					.showTextDocument(textDocument)
+					.then(changeTextEditor);
+			});
+		}
+
+		if (activeTextEditor?.document.uri.toString() === stringUri) {
+			const position = new Position(
+				message.jobOutput.position[0],
+				message.jobOutput.position[1],
+			);
+
+			const selection = new Selection(position, position);
+
+			activeTextEditor.selections = [selection];
+
+			activeTextEditor.revealRange(
+				new Range(position, position),
+				TextEditorRevealType.AtTop,
+			);
+		}
+
+        const allTextDocuments = textEditors
+			.map(({ document }) => document)
+			.concat(textDocuments);
     }
 }
